+++ built_value/lib/src/num_serializer.dart
@@ -4,6 +4,7 @@
 
 import 'package:built_collection/built_collection.dart';
 import 'package:built_value/serializer.dart';
+import 'package:built_value/src/double_serializer.dart';
 
 class NumSerializer implements PrimitiveSerializer<num> {
   final bool structured = false;
@@ -15,12 +16,28 @@ class NumSerializer implements PrimitiveSerializer<num> {
   @override
   Object serialize(Serializers serializers, num number,
       {FullType specifiedType: FullType.unspecified}) {
-    return number;
+    if (number.isNaN) {
+      return DoubleSerializer.NAN;
+    } else if (number.isInfinite) {
+      return number.isNegative
+          ? DoubleSerializer.NEGATIVE_INFINITY
+          : DoubleSerializer.INFINITY;
+    } else {
+      return number;
+    }
   }
 
   @override
   num deserialize(Serializers serializers, Object serialized,
       {FullType specifiedType: FullType.unspecified}) {
-    return serialized as num;
+    if (serialized == DoubleSerializer.NAN) {
+      return double.NAN;
+    } else if (serialized == DoubleSerializer.NEGATIVE_INFINITY) {
+      return double.NEGATIVE_INFINITY;
+    } else if (serialized == DoubleSerializer.INFINITY) {
+      return double.INFINITY;
+    } else {
+      return (serialized as num).toDouble();
+    }
   }
 }